Separating server unit tests into two files (#645)
